feat: migrate utils to seprate file
